Adds necessary capabilities to scamper binary
We would like to run all M-Lab container processes as a non-root user.
It turns out that the scamper binary requires a good number of
privileged capabilities. This commit adds those capabilities to the
scamper binary so that the container can run as non-root. This same
thing was done for scamper in the M-Lab traceroute-caller sidecar:

https://github.com/m-lab/traceroute-caller/pull/162